fixes #217: fixes up problems with Spring's ObjectMapper being overwritten
Now changes the configuration to use a SqsListenerObjectMapperSupplier
instead of doing a @ConditionalOnMissingBean for the ObjectMapper. The
reason is that the auto configuration for this library was running before
the Jackson auto configuration, resulting in it supplying an empty
ObjectMapper and thus removing the default implementation.

By using this supplier method, it will use the default Spring Boot
ObjectMapper if supplied, otherwise it will define its own custom empty
ObjectMapper.