fixes #223: fix shutdown race condition
This occurs if the broker thread had not been run before the container has
been requested to be shutdown. The result would be that the CompletableFuture
that indicates that the broker thread is finished would never be resolved
so we would sit here and wait forever.

This changes it to use the ExecutorService to shutdown all the threads.