refs #227 refs #228: Includes pom.xml in Jars, fixes gradle consumer
Add manual configuration for including the pom.xml and pom.properties
file in the Jar META-INF folder.

Gradle projects were failing to use the library unless they explicitly
defined the AWS BOM. This was not the case with Maven users. The cause
was the BOM being an `implementation` and not an `api` dependency.