refs #278: update formatting to use prettier instead of Checkstyle (#279)
This allows for auto formatting of code instead of running checkstyle linting checks and then manually fixing any problems.  The formatting rules are okay and there are a couple of places that I don't like, for example lambdas and functional chains (e.g. builders, streams, etc), but what are you going to do.